Cambie la ruta que renderiza el handlebars de search al archivo htmlRoutes.js para dejar unicamente las que contestan con json en el apiRoutes.js
